docs: adds selectors.js file
